+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - ✨(back) add ODT parsing support
 - ✨(back) add self-documentation tool
 - ✅(front) add tests for SourceItem component
+- ✨(back) de-index collections after inactivity 
 
 ### Changed